+++ addons/realisticnames/CfgWeapons.hpp
@@ -2,6 +2,7 @@ class Mode_SemiAuto;
 class Mode_FullAuto;
 
 class CfgWeapons {
+    #include "Attachments.hpp"
     // assault rifles
 
     // MX
@@ -707,175 +708,19 @@ class CfgWeapons {
         };
     };
 
-    //attachments
-
-    class ItemCore;
-
-    class acc_flashlight: ItemCore {
-        displayName = "UTG Defender 126";
-    };
-
-    class optic_Hamr: ItemCore {
-        displayName = CSTRING(optic_hamr);
-    };
-    class optic_Hamr_khk_F: optic_Hamr {
-        displayName = CSTRING(optic_hamr_khk);
-    };
-    class ACE_optic_Hamr_2D: optic_Hamr {
-        displayName = CSTRING(optic_hamr_2d);
-    };
-    class ACE_optic_Hamr_PIP: ACE_optic_Hamr_2D {
-        displayName = CSTRING(optic_hamr_pip);
-    };
-
-    class optic_Arco: ItemCore {
-        displayName = CSTRING(optic_arco);
-    };
-    class optic_Arco_blk_F: optic_Arco {
-        displayName = CSTRING(optic_arco_blk);
-    };
-    class optic_Arco_ghex_F: optic_Arco {
-        displayName = CSTRING(optic_arco_ghex);
-    };
-    class ACE_optic_Arco_2D: optic_Arco {
-        displayName = CSTRING(optic_arco_2d);
-    };
-    class ACE_optic_Arco_PIP: ACE_optic_Arco_2D {
-        displayName = CSTRING(optic_arco_pip);
-    };
-    class optic_Arco_lush_F: optic_Arco {
-        displayName = CSTRING(optic_arco_lush);
-    };
-    class optic_Arco_arid_F: optic_Arco {
-        displayName = CSTRING(optic_arco_arid);
-    };
-    class optic_Arco_AK_blk_F: optic_Arco_blk_F {
-        displayName = CSTRING(optic_arco_ak_blk);
-    };
-    class optic_Arco_AK_lush_F: optic_Arco_lush_F {
-        displayName = CSTRING(optic_arco_ak_lush);
-    };
-    class optic_Arco_AK_arid_F: optic_Arco_arid_F {
-        displayName = CSTRING(optic_arco_ak_arid);
-    };
-
-    class optic_ERCO_blk_f: optic_Arco {
-        displayName = CSTRING(optic_erco_blk);
-    };
-    class optic_ERCO_khk_f: optic_ERCO_blk_f {
-        displayName = CSTRING(optic_erco_khk);
-    };
-    class optic_ERCO_snd_f: optic_ERCO_blk_f {
-        displayName = CSTRING(optic_erco_snd);
-    };
-
-    class optic_LRPS: ItemCore {
-        displayName = CSTRING(optic_lrps);
-    };
-    class optic_LRPS_ghex_F: optic_LRPS {
-        displayName = CSTRING(optic_lrps_ghex);
-    };
-    class optic_LRPS_tna_F: optic_LRPS {
-        displayName = CSTRING(optic_lrps_tna);
-    };
-    class ACE_optic_LRPS_2D: optic_LRPS {
-        displayName = CSTRING(optic_lrps_2d);
-    };
-    class ACE_optic_LRPS_PIP: ACE_optic_LRPS_2D {
-        displayName = CSTRING(optic_lrps_pip);
-    };
-
-    class optic_AMS_base;
-    class optic_AMS: optic_AMS_base {
-        displayName = CSTRING(optic_ams);
-    };
-    class optic_AMS_khk: optic_AMS {
-        displayName = CSTRING(optic_ams_khk);
-    };
-    class optic_AMS_snd: optic_AMS {
-        displayName = CSTRING(optic_ams_snd);
-    };
-
-    class optic_KHS_base;
-    class optic_KHS_blk: optic_KHS_base {
-        displayName = CSTRING(optic_khs_blk);
-    };
-    class optic_KHS_hex: optic_KHS_blk {
-        displayName = CSTRING(optic_khs_hex);
-    };
-    class optic_KHS_old: ItemCore {
-        displayName = CSTRING(optic_khs_old);
-    };
-    class optic_KHS_tan: optic_KHS_blk {
-        displayName = CSTRING(optic_khs_tan);
-    };
-
-    class optic_DMS: ItemCore {
-        displayName = CSTRING(optic_dms);
-    };
-    class optic_DMS_ghex_F: optic_DMS {
-        displayName = CSTRING(optic_dms_ghex);
-    };
-    class optic_DMS_weathered_F: optic_DMS {
-        displayName = CSTRING(optic_dms_weathered);
-    };
-    class optic_DMS_weathered_Kir_F: optic_DMS_weathered_F {
-        displayName = CSTRING(optic_dms_weathered_kir);
-    };
-
-    class optic_holosight: ItemCore {
-        displayName = CSTRING(optic_holosight);
-    };
-    class optic_Holosight_blk_F: optic_holosight {
-        displayName = CSTRING(optic_holosight_blk);
-    };
-    class optic_Holosight_khk_F: optic_holosight {
-        displayName = CSTRING(optic_holosight_khk);
-    };
-    class optic_Holosight_lush_F: optic_holosight {
-        displayName = CSTRING(optic_holosight_lush);
-    };
-    class optic_Holosight_arid_F: optic_holosight {
-        displayName = CSTRING(optic_holosight_arid);
-    };
-    class optic_Holosight_smg: ItemCore {
-        displayName = CSTRING(optic_holosight_smg);
-    };
-    class optic_Holosight_smg_blk_F: optic_Holosight_smg {
-        displayName = CSTRING(optic_holosight_smg_blk);
-    };
-    class optic_Holosight_smg_khk_F: optic_Holosight_smg {
-        displayName = CSTRING(optic_holosight_smg_khk);
-    };
-
-    class optic_MRCO: ItemCore {
-        displayName = CSTRING(optic_MRCO);
-    };
-    class ACE_optic_MRCO_2D: optic_MRCO {
-        displayName = CSTRING(optic_MRCO_2d);
-    };
-    class ACE_optic_MRCO_PIP: ACE_optic_MRCO_2D {
-        displayName = CSTRING(optic_MRCO_pip);
-    };
-
-    class optic_Yorris: ItemCore {
-        displayName = CSTRING(optic_Yorris);
-    };
-
-    class optic_ACO: ItemCore {
-        displayName = CSTRING(optic_ACO);
-    };
-    class optic_ACO_grn: ItemCore {
-        displayName = CSTRING(optic_ACO_grn);
-    };
-    class optic_ACO_smg: ItemCore {
-        displayName = CSTRING(optic_ACO_smg);
-    };
-    class optic_ACO_grn_smg: ItemCore {
-        displayName = CSTRING(optic_ACO_grn_smg);
-    };
-
     // APEX/Tanoa
+
+    // Type 115
+    class arifle_ARX_base_F;
+    class arifle_ARX_blk_F: arifle_ARX_base_F {
+        displayName = CSTRING(arifle_arx_blk_Name);
+    };
+    class arifle_ARX_ghex_F: arifle_ARX_base_F {
+        displayName = CSTRING(arifle_arx_ghex_Name);
+    };
+    class arifle_ARX_hex_F: arifle_ARX_base_F {
+        displayName = CSTRING(arifle_arx_hex_Name);
+    };
 
     // QBZ-95 and variants
     class arifle_CTAR_base_F;
@@ -1017,6 +862,16 @@ class CfgWeapons {
     };
 
     // Contact/Livonia
+
+    // CZ 581 Shotgun
+    class sgun_HunterShotgun_01_base_F;
+    class sgun_HunterShotgun_01_F: sgun_HunterShotgun_01_base_F {
+        displayName = CSTRING(sgun_huntershotgun_01_Name);
+    };
+    class sgun_HunterShotgun_01_sawedoff_base_F;
+    class sgun_HunterShotgun_01_sawedoff_F: sgun_HunterShotgun_01_base_F {
+        displayName = CSTRING(sgun_huntershotgun_sawedoff_01_Name);
+    };
 
     // FNX-45 (Green)
     class hgun_Pistol_heavy_01_green_F: hgun_Pistol_heavy_01_F {
