+++ addons/realisticnames/stringtable.xml
@@ -698,6 +698,9 @@
             <Chinesesimp>"卡玛兹"（医疗）</Chinesesimp>
             <Turkish>KamAZ Medikal</Turkish>
         </Key>
+        <Key ID="STR_ACE_RealisticNames_Truck_02_water_Name">
+            <English>KamAZ Water</English>
+        </Key>
         <Key ID="STR_ACE_RealisticNames_Truck_02_MRL_Name">
             <English>KamAZ MRL</English>
             <German>KamAS MRL</German>
@@ -1654,6 +1657,12 @@
             <Korean>FNX-45 택티컬</Korean>
             <Chinese>FNX-45戰術型手槍</Chinese>
             <Chinesesimp>FNX-45 战术型</Chinesesimp>
+        </Key>
+        <Key ID="STR_ACE_RealisticNames_sgun_huntershotgun_01_Name">
+            <English>CZ 581</English>
+        </Key>
+        <Key ID="STR_ACE_RealisticNames_sgun_huntershotgun_sawedoff_01_Name">
+            <English>CZ 581 (Sawed-Off)</English>
         </Key>
         <Key ID="STR_ACE_RealisticNames_hgun_Pistol_heavy_01_green_Name">
             <English>FNX-45 Tactical (Green)</English>
@@ -3056,6 +3065,15 @@
             <Chinesesimp>"柏拉格" 战斗无人机</Chinesesimp>
             <Turkish>Burraq UCAV</Turkish>
         </Key>
+        <Key ID="STR_ACE_RealisticNames_arifle_arx_blk_Name">
+            <English>Type 115 (Black)</English>
+        </Key>
+        <Key ID="STR_ACE_RealisticNames_arifle_arx_ghex_Name">
+            <English>Type 115 (Green Hex)</English>
+        </Key>
+        <Key ID="STR_ACE_RealisticNames_arifle_arx_hex_Name">
+            <English>Type 115 (Hex)</English>
+        </Key>
         <Key ID="STR_ACE_RealisticNames_arifle_CTAR_blk">
             <English>QBZ-95-1 (Black)</English>
             <Czech>QBZ-95-1 (Černá)</Czech>
@@ -3839,6 +3857,15 @@
             <Turkish>Wiesel 2 RFCV (Radar)</Turkish>
             <Spanish>Wiesel 2 RFCV (Radar)</Spanish>
             <Korean>비젤 2 RFCV (레이더)</Korean>
+        </Key>
+        <Key ID="STR_ACE_RealisticNames_flashlight_Name">
+            <English>UTG Defender 126</English>
+        </Key>
+        <Key ID="STR_ACE_RealisticNames_optic_mrd_Name">
+            <English>EOTech MRDS</English>
+        </Key>
+        <Key ID="STR_ACE_RealisticNames_optic_mrd_black_Name">
+            <English>EOTech MRDS (Black)</English>
         </Key>
         <Key ID="STR_ACE_RealisticNames_optic_hamr">
             <English>Leupold Mark 4 HAMR</English>
